Create a utilities file
